Price fields are now "advanced" fields. Set values after fields added
The price fields were moved from their former template to commerce-provider specific templates. Those fields are added/displayed a tad later than when we were trying to set the values of the price fields. This change relocates the val() setting so the fields contain the proper data.

See: https://central.tri.be/issues/41003