Stop using the GLOB_BRACE flag
The `GLOB_BRACE` flag for `glob()` is not available on all systems. Most notably it's not supported on Solaris and Alpine (a quite popular lightweight linux distribution used mostly for running docker containers). On this PR I propose a way of achieving the same result which doesn't use such flag and hence improves compatibility.