Update tickets.php for Bug #81967
Similar to bug #82684, stock for tickets using global stock was being counted with each iteration, causing a multiple of the global stock to be returned for the count. Added in condition so that the ticket is "ignored" during the loop if it's using global stock, then the global stock is added to the total count at the end.