fix(Main): fix autoloader related issues
This fix makes sure that, even when the plugin is not loaded due to a missing requirement or
incompatibility, its source paths are registered, appending them, in the Autoloader. This makes it
so that the paths registered by the plugin owning Common will be searced **first** using the "safe"
version of the classes; if the classes are not found in the version of Common owned by the plugin
loading it then the additional plugin paths will be searched avoiding fatals caused by some plugin
functionalities relying on some classes that do not exist at all in the loaded (and probably older)
version of Common.

re https://central.tri.be/issues/127173

related to: https://github.com/moderntribe/the-events-calendar/pull/2493