fix(ticket header image): Allow adding ticket header image on non-event posts.
The block editor always used the event REST API for adding ticket(and RSVP) images - this fails if
we're not editing an event. This code changes it to detect the post type and use the appropriate
endpoint (`tribe_events` for events, `posts` otherwise).

Code changes are only to update and delete for both RSVP and Ticket.

:ticket: [ET-54]