fix(RSVP stock): Correct front end form when one RSVP becomes sold out.
THe rsvp form had a variable carried through the loop to check for "product available" - but if the
last RSVP was sold out then it would be set to `false` - resulting in the meta row no showing up.
We've updated that to rely on _two_ variables - one for the specific rsvp and one for all rsvps.

:movie_camera: http://p.tri.be/iRhxoL

:ticket: [ET-603]

BTW: 603 is the area code for NH :tada: