fix(global stock): Ensure we update the event when we add our first global stock ticket.
In some cases, adding a global stock ticket did not update the event meta that indicated it was
using global stock. Now, in the ticket save() function where we deal with global stock, if we have
not enabled global stock already, we make sure to update the event to indicate it is enabled.

:ticket: [ET-614]