[ETP-487] [ETP-517] Add function to get active providers for a post
[ETP-487]
[ETP-517]

This PR addresses a need by ET+ to only get providers for a post that are active. Other functions return the current provider set for a post but does not take into account whether it's active and available at the code level.

This PR also fixes a problem with the tribe-tickets-loader asset not being registered when it's needed by ET+ even when new rsvp/ticket views are not enabled.