refactor(modules,resources) rec. ev. tickets
Modify the logic that will prevent or allow tickets to be added to
recurring events, or events to be made recurrin when they already have
tickets, to read the setting from the backend, by means of teh
`tec-no-tickets-on-recurring` body class.
This will restore functionality to the non-CT1 version of the plugin.