fix(Tickets) attempt to repair the provider on fetch
For reasons that might be due to a past bug, some Events will have a
provider set that does not match that of the actual tickets attached
to the Event; e.g. the Event might have the ticket provider set to Tribe
Commerce while having WooCommerce tickets attached to it.

This draft fix will attempt to repair the provider, reading what is in
the database, once a week and only if no tickets were found.

The code in this commit is NOT meant to be final in implementation and
location, but to be used as a starting point.