feat(FT) filter labels during panel rendering
This commit updates the existing code to keep handling the `default`
ticket value but start handling different ticket types when provided.
The main modifications happen on the main main Tickets JS file now
always pulling the "New Ticket" panel from the backend and passing, in
the context of that operation, the ticket type to the backend.

The ticket type is then used from the backend to render panels on a
per-ticket-type basis.