Change implode order from legacy signature
This plugin sets its minimum PHP version as 7.4 in the readme.txt.

As of PHP 7.4 `implode()` changes the order of its arguments from 'Array, Separator' to 'Separator, 'Array'.

As of PHP 8.0 the legacy order support is removed.

In preparation for a minimum version of PHP 8.0 this would fix the removed error while being compatible with the plugins current minimal version.