fix(Blocks) ensure tribe/tickets-iten can be inserted
Since the code is not using the store provided by WordPress, there is a chance the logic to create the missing tickets blocks might
run while, or before, the WordPress store is initialized.
In the context of the `insertBlock` call the Block Editor code will run a check to see if the block to insert, `tribe/tickets-item`
in this case, can be inserted in the parent block (i.e. `tribe/tickets`); that check is based upon a state internal to the WordPress
store that might not be, at this stage, defined.
If the check turns out a falsy value, it will fail silently.
This commit ensures the `allowedBlocks` property of the `tribe/tickets` block is correctly set before trying to insert the
`tribe/tickets-item` block in it.