fix(FT) ensure only SP Attendee candidate checks-in once This updates the code taking care of cloning and checking in Series Pass Attendees to make sure candidates can only be checked in once.
Previously, cloned Attendees would, from the context of a check-in made with QR code, be incorrectly checked for already checked-in status.
If the Series Pass Attendee has an ID of 23, and the clone has an ID of 89, then the check for already checked-in status would be done,
incorrectly, on the Series Pass Attendee (never checked-in) and not on the cloned Attendee.