fix(FT) include series pass dates in PDF
Rework the logic in the method to include the series pass dates in the email to have them included in the PDF attachment as well.
When rendering the HTML for the PDF pass, the template input to the method will not be the Email one, but the Wallet Plus one.
For this reason, the controller is not using the input template, but hydrating its own, dedicated, instance.