refactor(Block Editor) filter the capacity form rendering
Refactor the capacity component to allow filtering the function
rendering the capacity form.

By default, the component will render a form that will be different
depending on Event Ticket Plus being active or not, but the filter
allows for plugins to jump into the form rendering and change it.

Returning an empty value will not render the form at all.