fix(Seating) service message queue processing
The service will send messages in order to sync with the site.

The previous message queue implementation relied on the actions
associated to the messages being all sync, thus processing them as soon
as they arrived it made sense.

A sync message processing system is not able, though, to process sync
and `async` handlers correctly though, as illustrated by this case:

1. The user edits a layout and clicks save.
2. The service sends the `LAYOUT_CREATED_UPDATED` message
3. The site message processing system fires the handler starting an AJAX
   request to the backend to trigger invalidation of the caches.
   This call takes time and is done by an `async` handler waiting on
   the `fetch` response with `await fetch( ... )`
3. The service sends the `GO_TO_LAYOUTS_HOME` message
4. The site immediately calles the sync handler that will redirect the
   user to the maps home

Depending on the situation, 3 and 4 might happen before the AJAX-calling
function in 2 is done, or even started. This will cause the effect of 2
(cache invalidation) to never happen.

To fix this I've rewritten the message processing sytem to treat **all**
message handlers are `async` and wait for one to be done to call the
next. This creates a queue that is processed in asynchronous mode.
The implementation relies on the fact that one can `wait` on **any**
function, not just `async` functions.

This code will log `23`:

```javascript
const getNum = () => 23;
const number = await getNum();
console.log(number);
```

Since I felt queasy setting up things without tests, I've extended the
current Jest tests to include the Seating feature in the
`tests/slr_jest` directory.