SLR - clear reservations on seat selection cancel
This updates the ticket block frontend code to make sure reservations
will be cancelled when a user cancels the seat selection action by
either:
* pressing Esc on the keyboard
* clicking outside the modal main area
* clicking the modal X button
* clicking the Cancel button in the modal body

The fix comes with Jest tests.

Along with the fix some smaller refactorings along the way.

Note: there are other instances where the reservations should be
canceled at a later stage: when the user cancels the checkout, the
user cancels during the Attendee registration or when the user lets the
timer expire. Those instances will be handled in another PR.