fix(Seating) seat assignment to Attendees
This updates the code around seat assigment to Attendees to take into
account different stock reduction flows, hence Attendee generation
flows.

Depending on when the stock should be reduced and the Attendee
generated, the data required to assign Attendees might still be in the
session or not.

This updates the code to handle both instances and warmup caches used by
the Cart to assign Attendees in either cases.