Ensure that only float values are passed to number_format
PHP 8.0+ throws fatals when non-float values are passed to number_format, which can occur from time to time. A customer reported this issue happening and he has to modify the code each time a release happens. This should fix that for him!

:ticket: [ETP-962]

The passing tests will act as the artifact for this PR as the change is incredibly low-level.