refactor(Blocks/Ticket) remove filtering of edit comp
The intent behind the original filtering seemed to be to remove the
tickets from the post/event when the Ticket item is removed. This is not
what is currently happening, so this code seems to not fulfill its
intent, and it's generating a fatal on the first click of the "Add a
Ticket" button in the Tickets (plural) block.