Deprecate the old landing page and redirect to the new one.
This deprecates the Main.php instance of Tribe__Admin__Activation_Page and the function that instantiates it.
Instead, it leverages the new Onboarding wizard controller to handle the redirect.

The redirect happens immediately on a standalone ET install.
After a bulk activation of ET and other plugins, it will happen on the next ET admin page load.

[ET-2358] [ET-2359]