fix(Views v2) do not cache providers w/ tickets
The ticket providers (e.g. Tickets Commerce or WooCommerce) are
singletons that should not be cached along with the Ticket Information.
While the JSON packer **can** cache them, it should not to avoid
duplicated data information. Unsetting the `prvovider` property of the
Tickets before the packing will set it to `null` and will fetch the
provider singleton instance correctly usign the logic in the
`get_provider` method, based on the `provider_class` property.