fix(Classy) hard-code the TEC store name
The `@tec/events/classy/constants` import will resolve to the
`window.tec.events.classy.constants` object that might not yet be
defined at module-load time.
Imports are resolved immediately and, depending on the loading order of
the JS files, the TEC Classy object might not be defined yet when the ET
modules try to access it.