fix: remove meta key that would require migration
While the use of the `_stock_status` meta key makes sense from the
performance point of view, it would excluded from queries the RSVP
Tickets that do not have it. This commit updates the code to bring the
refactored repository code in line with the previous version of the RSVP
code.