refactor: port over most of classic editor logic
Continue the porting, from the feature/rsvp-merge-style branch, of the
Classic Editor code, scripts and style to get to feature parity with the
previous implementation. The location and structure of the code reflects
the least amount of effort required to move the functionality over.
Following changes will deal with correct positioning of the code in the
correct locations and refactorings.